requirements.txt
annotated-types==0.7.0
bandit==1.8.3
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
docstring_parser==0.16
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mailman==1.1.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.88.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-cloud-vision==3.8.1
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
packaging==24.2
pbr==6.1.1
pdf2image==1.17.0
pdftoimg==0.1.1
pillow==11.0.0
pinecone==5.4.1
pinecone-plugin-inference==3.0.0
pinecone-plugin-interface==0.0.7
proto-plus==1.25.0
protobuf==5.29.0
protoc-gen-openapiv2==0.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.10.0
pymongo==4.10.1
PyMuPDF==1.24.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9
shapely==2.0.6
six==1.16.0
stevedore==5.4.1
stripe==11.3.0
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
